Added error parser to exported API
