+++ src/api/CompileTools.js
@@ -4,7 +4,6 @@ const path = require(`path`);
 
 const gitExtension = vscode.extensions.getExtension(`vscode.git`).exports;
 
-const { default: IBMi } = require(`./IBMi`);
 const { GlobalConfiguration, ConnectionConfiguration } = require(`./Configuration`);
 const { CustomUI, Field } = require(`./CustomUI`);
 const { getEnvConfig } = require(`./local/env`);
@@ -12,9 +11,6 @@ const { getLocalActions, getiProjActions } = require(`./local/actions`);
 const { Deployment } = require(`./local/deployment`);
 const { parseErrors } = require(`./errors/handler`);
 const { default: Instance } = require(`./Instance`);
-
-const { Deployment } = require(`./local/deployment`);
-const { parseErrors } = require(`./errors/handler`);
 
 const diagnosticSeverity = {
   0: vscode.DiagnosticSeverity.Information,
