+++ src/extension.ts
@@ -11,6 +11,7 @@ import {ObjectBrowserProvider} from "./views/ConnectionBrowser";
 import IBMi from "./api/IBMi";
 import { ConnectionConfiguration } from "./api/Configuration";
 import { CodeForIBMi, ConnectionData } from "./typings";
+import { parseErrors } from "./api/errors/handler";
 
 export function activate(context: ExtensionContext): CodeForIBMi {
   // Use the console to output diagnostic information (console.log) and errors (console.error)
@@ -55,7 +56,7 @@ export function activate(context: ExtensionContext): CodeForIBMi {
     })
   )
 
-  return { instance, CustomUI, Field, baseContext: context };
+  return { instance, CustomUI, Field, baseContext: context, evfeventParser: parseErrors };
 }
 
 // this method is called when your extension is deactivated
