+++ src/typings.ts
@@ -6,6 +6,7 @@ export interface CodeForIBMi {
   baseContext: ExtensionContext,
   CustomUI: object, //CustomUI: typeof CustomUI
   Field: object //Field: typeof Field;
+  evfeventParser: (lines: string[]) => Map<string, FileError[]>
 }
 
 export interface StandardIO {
@@ -101,4 +102,13 @@ export interface IBMiError{
 export interface Filter{
   library: string,
   filter: string
+}
+
+export interface FileError {
+  sev: number
+  linenum: number
+  column: number
+  toColumn: number
+  text: string
+  code: string
 }
